dissolve the planning scc with three seam breaks
